# backend/tests/contracts/openapi.corvus-locomo.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Agentic institutional memory for coding harnesses. Situated episodes with outcomes, evidence-gated lessons, temporal supersession.",
    "title": "Corvus LoCoMo Eval",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/janitor/as-of/{neuron_id}": {
      "get": {
        "description": "What did we believe about this memory row at datetime `at`?\n\nReverts change-log events newer than `at` (newest first) to\nreconstruct the row's logged fields as they stood at that moment.",
        "operationId": "memory_as_of_janitor_as_of__neuron_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "neuron_id",
            "required": true,
            "schema": {
              "title": "Neuron Id",
              "type": "integer"
            }
          },
          {
            "description": "ISO datetime — reconstruct belief state then",
            "in": "query",
            "name": "at",
            "required": true,
            "schema": {
              "description": "ISO datetime — reconstruct belief state then",
              "title": "At",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Memory As Of",
        "tags": [
          "memory"
        ]
      }
    },
...
    "/janitor/history/{neuron_id}": {
      "get": {
        "description": "Temporal view of one memory row (kill-temporal-kg parity).\n\nReturns the row's change log plus its predecessors — rows it\nsuperseded, each with a validity window (created_at .. the\nchanged_at of the supersession event).",
        "operationId": "memory_history_janitor_history__neuron_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "neuron_id",
            "required": true,
            "schema": {
              "title": "Neuron Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Memory History",
        "tags": [
          "memory"
        ]
      }
    },
    "/janitor/run": {
      "post": {
        "description": "Run the selected janitor passes and return the combined report.",
        "operationId": "janitor_run_janitor_run_post",
        "parameters": [
          {
            "in": "query",
            "name": "consolidation",
            "required": false,
            "schema": {
              "default": true,
              "title": "Consolidation",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "staleness",
            "required": false,
            "schema": {
              "default": true,
              "title": "Staleness",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "decay",
            "required": false,
            "schema": {
              "default": true,
              "title": "Decay",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "lint",
            "required": false,
            "schema": {
              "default": true,
              "title": "Lint",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "plasticity",
            "required": false,
            "schema": {
              "default": true,
              "title": "Plasticity",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "max_pairs",
            "required": false,
            "schema": {
              "default": 40,
              "maximum": 200,
              "minimum": 1,
              "title": "Max Pairs",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Janitor Run",
        "tags": [
          "memory"
        ]
      }
    },
...
  }
}